Fix beds (#163)
Visually janky, but functional beds. Preserves spawn points and allows sleeping. Bed distance check has been fixed so that players can use beds on ships. The lie down animation is untouched, as that's part of player rendering, so the player will not properly lie down on the bed, but the bed is now usable for sleeping and setting spawn.

Co-authored-by: Rubydesic <rubydesic@gmail.com>