Fix for some entities crashing when getting added to a constructed ship.
This change adds a check against the entity's renderer to see if it is eligible for rendering before it calls the render function. This prevents crashes from occurring where an entity may not be ready to render yet, i.e. in the case of Mekanism's Universal Cables.

Also changed the way the java version is set for the project, so developers that have multiple java versions installed can still use this project without needing to reset their java home to java 17.